adapter new type promotion rule for Paddle 2.6 (#455)
* adapter new type promotion rule for Paddle 2.6

* fix code style

+++ paddle3d/models/detection/smoke/smoke_coder.py
@@ -133,9 +133,8 @@ class SMOKECoder(paddle.nn.Layer):
         points = paddle.reshape(points, (-1, 2))
         assert points.shape[0] == N
 
-        # int + float -> int, but float + int -> float
         # proj_points = points + points_offset
-        proj_points = points_offset + points
+        proj_points = points_offset + points.astype(points_offset.dtype)
 
         # transform project points in homogeneous form.
         proj_points_extend = paddle.concat(
